refactor Adams Bohart results to new component
